+++ src/main/java/kitchenpos/application/MenuService.java
@@ -4,11 +4,11 @@ import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
-import kitchenpos.menu.Menu;
-import kitchenpos.menu.MenuGroupRepository;
-import kitchenpos.menu.MenuProduct;
-import kitchenpos.menu.MenuProductRepository;
-import kitchenpos.menu.MenuRepository;
+import kitchenpos.menu.domain.Menu;
+import kitchenpos.menu.domain.MenuGroupRepository;
+import kitchenpos.menu.domain.MenuProduct;
+import kitchenpos.menu.domain.MenuProductRepository;
+import kitchenpos.menu.domain.MenuRepository;
 import kitchenpos.product.domain.Product;
 import kitchenpos.product.domain.ProductRepository;
 import org.springframework.stereotype.Service;
