+++ src/main/java/kitchenpos/product/domain/Product.java
@@ -8,7 +8,7 @@ import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.OneToMany;
-import kitchenpos.menu.MenuProduct;
+import kitchenpos.menu.domain.MenuProduct;
 
 @Entity
 public class Product {
