+++ src/test/java/kitchenpos/application/MenuServiceTest.java
@@ -8,11 +8,11 @@ import static org.mockito.BDDMockito.given;
 import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
-import kitchenpos.menu.Menu;
-import kitchenpos.menu.MenuGroupRepository;
-import kitchenpos.menu.MenuProduct;
-import kitchenpos.menu.MenuProductRepository;
-import kitchenpos.menu.MenuRepository;
+import kitchenpos.menu.domain.Menu;
+import kitchenpos.menu.domain.MenuGroupRepository;
+import kitchenpos.menu.domain.MenuProduct;
+import kitchenpos.menu.domain.MenuProductRepository;
+import kitchenpos.menu.domain.MenuRepository;
 import kitchenpos.product.domain.Product;
 import kitchenpos.product.domain.ProductRepository;
 import org.assertj.core.util.Lists;
